We can play whole game!!!
But, there are things still to be done:
(Jack)
Ensure the bottom of the red felt is the bottom of the screen
Lower the user's cards on the red felt
Lower the top of the red felt
Make an image button to show the top card to User. It will go where the user's cards move to when they are played. Make it invisible when topCardTurnedDown = true is set or when pickItUP() is called.
Add "waits" after every AI play (2 - 5 seconds).
Add text box that give User feed back like "Choose a card to discard" when player is dealer and has to pickItUp and also for moments like "Your teammate passes."/"Player left of you called trump as HEARTS".
Do any other TODOs left in GameScreen.java

(Everyone else)
Clean up unused code
comment, comment, comment
test, test, test